tweak: don't reindent after killing lines
Killing lines is not much different from killing words/chars. Sometimes
we want to reindent after killing a word, but sometimes we don't, since
we may just want to fix a typo in that word. Same situations happen to
killing lines, so it's better to not automatically reindent.

Besides, by getting rid of this, we keep the implementation simple, and
this matters as Puni encourages the user to define their own commands,
and simpler implementation serves as a better reference.